python/requirements.txt
pyserial>=3.5
pandas>=2.0
tensorflow>=2.16
numpy>=1.24
scikit-learn>=1.0
